🐛🎨 Renaming and fixes
