+++ .mocharc.json
@@ -3,7 +3,7 @@
     "@babel/register",
     "@babel/polyfill",
     "jsdom-global/register",
-    " mock-local-storage",
+    "mock-local-storage",
     "./test/setup.js"
   ]
 }
