+++ README.md
@@ -91,7 +91,7 @@ $ yarn add beautiful-react-hooks
 * [useInterval](docs/useInterval.md)
 * [useDebouncedFn](docs/useDebouncedFn.md)
 * [useThrottledFn](docs/useThrottledFn.md)
-* [usePersistState](docs/usePersistState.md)
+* [useLocalStorage](docs/useLocalStorage.md)
 
 <div>
   <p align="center">
