+++ src/useLocalStorage.spec.js
@@ -1,9 +1,9 @@
 import React, { useEffect } from 'react';
 import { render, cleanup as cleanupReact } from '@testing-library/react';
 import { cleanup as cleanupHooks, renderHook } from '@testing-library/react-hooks';
-import usePersistState from './usePersistState';
+import useLocalStorage from './useLocalStorage';
 
-describe('usePersistState', () => {
+describe('useLocalStorage', () => {
   beforeEach(() => {
     cleanupHooks();
     cleanupReact();
@@ -11,10 +11,19 @@ describe('usePersistState', () => {
   });
 
   it('should be a function', () => {
-    expect(usePersistState).to.be.a('function');
+    expect(useLocalStorage).to.be.a('function');
   });
+
+  it('should return null when no default value defined', () => {
+    const { result, rerender } = renderHook(() => useLocalStorage('storageKey_1'));
+    const [value] = result.current;
+    rerender();
+    expect(value).to.equal(null);
+  });
+
+
   it('should return default value', () => {
-    const { result, rerender } = renderHook(() => usePersistState('storageKey', 100));
+    const { result, rerender } = renderHook(() => useLocalStorage('storageKey_2', 100));
     const [value] = result.current;
     rerender();
 
@@ -25,7 +34,7 @@ describe('usePersistState', () => {
     const TestComponent = (props) => {
       // eslint-disable-next-line react/prop-types
       const { newValue } = props;
-      const [value, setValue] = usePersistState('storageKey', 100);
+      const [value, setValue] = useLocalStorage('storageKey_2', 100);
 
       const setNewState = (v) => {
         setValue(v);
@@ -38,8 +47,7 @@ describe('usePersistState', () => {
       return <p>{value}</p>;
     };
 
-    const { container, rerender } = render(<TestComponent newValue={200} />);
-    rerender(<TestComponent newValue={300} />);
+    const { container } = render(<TestComponent newValue={200} />);
 
     expect(container.querySelector('p').innerHTML).to.equal('200');
   });
